+++ src/pipex.c
@@ -6,17 +6,17 @@
 /*   By: luguimar <marvin@42.fr>                    +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2023/09/21 17:07:51 by luguimar          #+#    #+#             */
-/*   Updated: 2024/04/17 09:36:00 by luguimar         ###   ########.fr       */
+/*   Updated: 2024/04/17 13:08:21 by luguimar         ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
 #include "../includes/minishell.h"
 
-static void	exec_command(char *path, char **envp, char **args, int isparent)
+static void	exec_command(char *path, t_shell *shell, char **args, int isparent)
 {
-	/*if (exec_builtin(args, envp))
-		return ;*/
-	if (!path || !envp || !args)
+	if (exec_builtin(args, shell))
+		return ;
+	if (!path || !shell->env_array || !args)
 	{
 		while (--isparent)
 			wait(NULL);
@@ -30,7 +30,7 @@ static void	exec_command(char *path, char **envp, char **args, int isparent)
 		free_array_of_strings(args);
 		exit(127);
 	}
-	execve(path, args, envp);
+	execve(path, args, shell->env_array);
 	if (isparent)
 		wait(NULL);
 	dup2(STDERR_FILENO, STDOUT_FILENO);
@@ -84,7 +84,7 @@ char	*get_right_path(char **cmd, char **envp, char *right_path)
 	return (NULL);
 }
 
-void	redirect_files(int i, char *argv[], char **envp)
+void	redirect_files(int i, char *argv[], t_shell *shell)
 {
 	int		cid;
 	int		pipefd[2];
@@ -98,11 +98,11 @@ void	redirect_files(int i, char *argv[], char **envp)
 	if (cid == 0)
 	{
 		args = ft_splitquote_nulls(argv[ft_abs_value(i)], ' ');
-		path = get_right_path(args, envp, path);
+		path = get_right_path(args, shell->env_array, path);
 		//if (i == 2 || i == -3)
 		//	check_error(access(argv[1], R_OK), argv[1], args, path);
 		dup2stdout(pipefd);
-		exec_command(path, envp, args, 1);
+		exec_command(path, shell, args, 1);
 	}
 	else if (cid == -1)
 		check_error(-1, "fork", args, path);
@@ -113,7 +113,7 @@ void	redirect_files(int i, char *argv[], char **envp)
 	}
 }
 
-int	pipex(int argc, char **argv, char **envp)
+int	pipex(int argc, char **argv, t_shell *shell)
 {
 	int		fd[2];
 	int		i;
@@ -123,10 +123,10 @@ int	pipex(int argc, char **argv, char **envp)
 	(void)fd;
 	i = -1;
 	path = NULL;
-	//dup2redirect(fd, argv, envp, i);
+	//dup2redirect(fd, argv, shell, i);
 	while (++i < argc - 1)
-		redirect_files(i, argv, envp);
-	args = last_one(argv, &path, envp, i);
-	exec_command(path, envp, args, 1);
+		redirect_files(i, argv, shell);
+	args = last_one(argv, &path, shell->env_array, i);
+	exec_command(path, shell, args, 1);
 	return (1);
 }
